1.0.5 automatically generated from 053b420

+++ src/main/CsharpDotNet2/IO/Swagger/Api/UsersApi.cs
@@ -20,7 +20,7 @@ namespace IO.Swagger.Api
         /// <returns></returns>
         void DeleteUser (string projectKey, string environmentKey, string userKey);
         /// <summary>
-        /// Search users in LaunchDarkly based on their last active date, or a search query. 
+        /// Search users in LaunchDarkly based on their last active date, or a search query. It should not be used to enumerate all users in LaunchDarkly- - use the List users API resource. 
         /// </summary>
         /// <param name="projectKey">The project key, used to tie the flags together under one project so they can be managed together.</param>
         /// <param name="environmentKey">The environment key</param>
@@ -29,7 +29,7 @@ namespace IO.Swagger.Api
         /// <param name="offset">Specifies the first item to return in the collection</param>
         /// <param name="after">A unix epoch time in milliseconds specifying the maximum last time a user requested a feature flag</param>
         /// <returns>Users</returns>
-        Users GetSearchUsers (string projectKey, string environmentKey, string q, decimal? limit, decimal? offset, decimal? after);
+        Users GetSearchUsers (string projectKey, string environmentKey, string q, decimal? limit, decimal? offset, long? after);
         /// <summary>
         /// Get a user by key. 
         /// </summary>
@@ -39,7 +39,7 @@ namespace IO.Swagger.Api
         /// <returns>User</returns>
         User GetUser (string projectKey, string environmentKey, string userKey);
         /// <summary>
-        /// List all users in the environment. 
+        /// List all users in the environment. Includes the total count of users. In each page, there will be up to &#39;limit&#39; users returned (default 20). This is useful for exporting all users in the system for further analysis. Paginated collections will include a next link containing a URL with the next set of elements in the collection. 
         /// </summary>
         /// <param name="projectKey">The project key, used to tie the flags together under one project so they can be managed together.</param>
         /// <param name="environmentKey">The environment key</param>
@@ -149,7 +149,7 @@ path = path.Replace("{" + "userKey" + "}", ApiClient.ParameterToString(userKey))
         }
     
         /// <summary>
-        /// Search users in LaunchDarkly based on their last active date, or a search query. 
+        /// Search users in LaunchDarkly based on their last active date, or a search query. It should not be used to enumerate all users in LaunchDarkly- - use the List users API resource. 
         /// </summary>
         /// <param name="projectKey">The project key, used to tie the flags together under one project so they can be managed together.</param> 
         /// <param name="environmentKey">The environment key</param> 
@@ -158,7 +158,7 @@ path = path.Replace("{" + "userKey" + "}", ApiClient.ParameterToString(userKey))
         /// <param name="offset">Specifies the first item to return in the collection</param> 
         /// <param name="after">A unix epoch time in milliseconds specifying the maximum last time a user requested a feature flag</param> 
         /// <returns>Users</returns>            
-        public Users GetSearchUsers (string projectKey, string environmentKey, string q, decimal? limit, decimal? offset, decimal? after)
+        public Users GetSearchUsers (string projectKey, string environmentKey, string q, decimal? limit, decimal? offset, long? after)
         {
             
             // verify the required parameter 'projectKey' is set
@@ -246,7 +246,7 @@ path = path.Replace("{" + "userKey" + "}", ApiClient.ParameterToString(userKey))
         }
     
         /// <summary>
-        /// List all users in the environment. 
+        /// List all users in the environment. Includes the total count of users. In each page, there will be up to &#39;limit&#39; users returned (default 20). This is useful for exporting all users in the system for further analysis. Paginated collections will include a next link containing a URL with the next set of elements in the collection. 
         /// </summary>
         /// <param name="projectKey">The project key, used to tie the flags together under one project so they can be managed together.</param> 
         /// <param name="environmentKey">The environment key</param>
